refactor(locale): move all translations into a locales folder so i18n ally will detect them
