--- src/translations/Chinese.ts
@@ -1,140 +1,0 @@
-import { TranslationObject } from './English'
-
-export const ChineseTranslations: TranslationObject = {
-  startupDescriptions: {
-    admin: {
-      password: '为 Procon 设置 RCON 密码',
-    },
-    vars: {
-      ranked: '启用排名',
-      serverName: '设置服务器名称',
-      gamePassword: '设置加入服务器的密码',
-      autoBalance: '启用自动平衡',
-      roundStartPlayerCount: '设置最小玩家以激活回合前/回合内',
-      roundRestartPlayerCount: '设置最小玩家数量以恢复热身',
-      roundLockdownCountdown: '设置回合前持续时间',
-      serverMessage: '设置服务器欢迎消息',
-      friendlyFire: '启用团队伤害',
-      maxPlayers: '设置最大玩家',
-      serverDescription: '设置服务器描述',
-      killCam: '启用击杀镜头',
-      miniMap: '启用小地图',
-      hud: '启用HUD',
-      crossHair: '为所有武器启用十字准线',
-      _3dSpotting: '在发现的敌人上启用3D标记',
-      miniMapSpotting: '在小地图上启用发现的敌人',
-      nameTag: '启用姓名标签',
-      _3pCam: '启用第三人称车辆摄像头',
-      regenerateHealth: '启用健康再生',
-      teamKillCountForKick: '设置一轮中允许的团队击杀次数',
-      teamKillValueForKick: '设置玩家在被踢出前允许的最大击杀值',
-      teamKillValuelncrease: '设置团队击杀的击杀值增加',
-      teamKillValueDecreasePerSecond: '设置每秒击杀值减少',
-      teamKillKickForBan: '设置导致永久禁赛的团队击杀次数',
-      idleTimeout: '设置空闲超时',
-      idleBanRounds: '设置应禁止多少轮空闲超时（如果有）',
-      vehicleSpawnAllowed: '启用车辆生成',
-      vehicleSpawnDelay: '设置车辆生成延迟比例因子',
-      soldierHealth: '设置士兵最大健康比例因子',
-      playerRespawnTime: '设置玩家重生时间尺度因子',
-      playerManDownTime: '设置玩家倒地时间比例因子',
-      bulletDamage: '设置子弹伤害比例因子',
-      gameModeCounter: '设置结束回合的票数比例因子',
-      onlySquadLeaderSpawn: '禁用小队队长以外的小队成员重生',
-      // unlockMode:  "将武器和装备设置为在未评级服务器上可用",
-    },
-    RM: {
-      setDevelopers: '设置开发人员列表，以空格分隔',
-      setAdmins: '设置管理员列表，以空格分隔',
-      setLightAdmins: '设置轻量级管理员列表，以空格分隔',
-      serverInfo: '服务器描述',
-      serverLicenseKey: 'RM 服务器许可证密钥',
-      ingameBanner: '图片链接',
-      pingLimitEnable: '激活 ping 限制',
-      pingLimitInMs: '时间（毫秒）',
-      autoPerfEnabled: '激活自动性能',
-      autoPerfMaxPlayers: '自动性能截止',
-      tempReservedSlotsEnabled: '启用自动重新加入',
-      tempReservedSlotsRejoinTime: '重新加入前多久被移除',
-      defaultPreRoundTime: '回合开始前的时间',
-      setAutoBalancer: '自动平衡',
-      battleCryLink: '链接到您的 battlecry JSON 端点',
-    },
-    vu: {
-      ColorCorrectionEnabled: '启用蓝色滤镜',
-      DesertingAllowed: '禁用越界',
-      DestructionEnabled: '启用破坏',
-      HighPerformanceReplication: '启用远处玩家的高性能更新',
-      ServerBanner: '设置指向 1378x162 jpg 的浏览器横幅 URL',
-      SetTeamTicketCount: '设置团队的团队票数。',
-      SquadSize: '设置每个小队的最大玩家数量',
-      SunFlareEnabled: '启用太阳耀斑',
-      SuppressionMultiplier: '设置压制强度 [0，无限]',
-      FriendlyFireSuppression: '启用对盟友的压制',
-      TimeScale: '减慢或加快游戏速度 [0.0，2.0]',
-      VehicleDisablingEnabled: '启用对车辆的移动命中',
-      HttpAssetUrl: '设置指向远程 [ui.vuic] 的 URL，而不是从游戏服务器下载',
-      DisablePreRound: '禁用等待玩家',
-      TeamActivatedMines: '启用地雷以触发盟友',
-      CorpseDamageEnabled: '启用尸体伤害以防止复活',
-    },
-    reservedSlots: {
-      add: '要添加到预留位置的玩家列表，以逗号分隔',
-    },
-  },
-  launchArgumentDescriptions: {
-    common: {
-      gamepath: '设置自定义 Battlefield 3 安装目录。',
-      perftrace: '启用性能日志记录到 perftrace-[server|client].csv',
-      env: '设置 Zeus 环境。默认为 prod。',
-      updateBranch: '设置更新分支。默认为 -env 所设置的任何值。',
-      tracedc: '在 VEXT 中启用 DataContainer Trace 使用，并在关卡销毁期间打印任何悬空 DC。',
-      cacert: '设置自定义 CA 证书包以用于 SSL 验证。',
-      activateWithOrigin:
-        '-activate -o_mail <email> -o_pass <pass> 使用指定的 Origin 凭据在当前计算机上激活 BF3。',
-      activateWithLSX:
-        '-activate -lsx 使用 EA app/Origin (LSX) 在当前计算机上激活 BF3，该应用程序/Origin (LSX) 必须正在运行并登录到拥有 BF3 的帐户。它还会打印从 EA 应用程序检索到的身份验证令牌。',
-      activateWithEaToken:
-        '-activate -ea_token <token> 使用提供的 EA 身份验证令牌在当前计算机上激活 BF3。可以使用上述命令从另一台计算机检索令牌。请记住，令牌会过期。',
-      console: '启用外部控制台窗口以进行调试日志记录。',
-      debuglog:
-        '启用日志记录。服务器日志位于服务器实例目录中。客户端日志位于 VU AppData 安装文件夹中。',
-      trace: '启用详细日志记录。',
-      vextdebug:
-        '设置 VEXT 远程调试主机：端口（目前不起作用）。它还可以防止服务器/客户端连接超时，并使发送到客户端的编译 VEXT 模块包含调试符号。',
-      vexttrace:
-        '启用 VEXT 执行跟踪。启用跟踪且 VU 崩溃时，每个加载的 VEXT 模块的最后一行执行内容将从崩溃对话框和提交的崩溃详细信息中提供。请记住，这可能会对性能产生不利影响。',
-    },
-    client: {
-      dwebui: '在 http://localhost:8884 启用 WebUI 调试。',
-      serverJoinString:
-        '使用给定的 GUID 和可选密码连接到服务器。可以用作加入的超链接。如果加入和观看字符串为空，VU Launcher 将自动在后端创建加入字符串。vu://join/4ffa57c1068540e59b9a25eb6b00dfcd/amazingpass',
-      serverSpectateString:
-        '使用给定的 GUID 和可选密码作为观看者连接到服务器。可以用作加入的超链接。如果加入和观看字符串为空，VU Launcher 将自动在后端创建加入字符串。vu://spectate/4ffa57c1068540e59b9a25eb6b00dfcd/amazingpass',
-      cefdebug:
-        '为 CEF for WebUI mods 启用详细调试日志记录。将在 VU 目录中创建一个 debug.log 文件（通常为 %LocalAppData%\\VeniceUnleashed\\client）。',
-      credentials: '-username <用户名> -password <密码> 使用指定的 VU 凭据登录到您的 VU 帐户。',
-      disableUiHwAcceleration:
-        '禁用 WebUI 的硬件加速。仅在启动 VU 出现黑色加载屏幕时使用。在未支持的操作系统上默认启用。',
-    },
-    server: {
-      high60: '将 VU 服务器频率设置为 60Hz。',
-      high120: '将 VU 服务器频率设置为 120Hz。',
-      headless: '以无头模式运行 VU 服务器（不创建任何窗口）。',
-      serverInstancePath: '为 VU 服务器设置服务器实例路径（存储服务器配置、日志和模块的位置）。',
-      highResTerrain: '启用高分辨率地形。用于将地图扩展到其原始游戏区域之外。',
-      disableTerrainInterpolation: '禁用不同地形 LOD 之间的插值。',
-      skipChecksum: '禁用客户端连接上的级别校验和验证。',
-      listen: '设置 VU 服务器应监听连接的主机和端口。默认为 0.0.0.0:25200',
-      mHarmonyPort: '设置 VU 服务器应监听 MonitoredHarmony 连接的端口。默认为 7948',
-      remoteAdminPort: '设置 VU 服务器应监听 RCON 连接的主机和端口。默认为 0.0.0.0:47200',
-      unlisted: '从服务器列表中隐藏服务器。加入需要 connect 控制台命令或 vu join 字符串。',
-      joinaddr: '设置客户端应连接到的 IPv4 地址。如果为空，后端将尝试自动检测服务器的 IP 地址。',
-      joinhost:
-        '设置客户端在连接时应使用的主机名。指定此主机名时，-joinaddr 无效，任何尝试连接的客户端都不会尝试执行任何 NAT 检测。',
-      noUpdate: '禁用自动更新。',
-      maxPlayers:
-        '设置最大玩家数 (vars.maxPlayers + vars.maxSpectators)。使用默认观众人数 24 时，您无法将 vars.maxPlayers 设置为高于 40。',
-    },
-  },
-}
