--- src/translations/manual-translation-helper.ts
@@ -1,13 +1,0 @@
-// this file is just to make it more convenient to batch translate
-
-const before = {
-	someStructure: {
-		abc: "blabla"
-	}
-}
-
-const after = {
-	someStructure: {
-		abc: "blabla"		
-	}
-}
